Move common helpers to src/ (#465)
