+++ src/cmds/core.rs
@@ -1,11 +1,11 @@
 use crate::cheatsh;
-use crate::common::clipboard;
-use crate::common::shell::{BashSpawnError, IS_FISH};
+use crate::clipboard;
 use crate::display;
 use crate::env_vars;
 use crate::fetcher::Fetcher;
 use crate::filesystem;
 use crate::finder::Finder;
+use crate::shell::{BashSpawnError, IS_FISH};
 use crate::structures::cheat::{Suggestion, VariableMap};
 use crate::structures::config::Action;
 use crate::structures::config::Config;
