--- src/common/mod.rs
@@ -1,8 +1,0 @@
-pub mod clipboard;
-pub mod file_issue;
-pub mod filesystem;
-pub mod git;
-pub mod hash;
-pub mod shell;
-pub mod terminal_width;
-pub mod url;
