+++ src/filesystem.rs
@@ -1,9 +1,7 @@
-pub use crate::common::filesystem::{
-    create_dir, exe_string, pathbuf_to_string, remove_dir, InvalidPath, UnreadableDir,
-};
 use crate::display::Writer;
 use crate::fetcher;
 pub use crate::fetcher::filesystem::{all_cheat_files, default_cheat_pathbuf, read_all};
+pub use crate::fs::{create_dir, exe_string, pathbuf_to_string, remove_dir, InvalidPath, UnreadableDir};
 use crate::structures::cheat::VariableMap;
 use anyhow::Error;
