+++ src/structures/cheat.rs
@@ -1,4 +1,4 @@
-use crate::common::hash::fnv;
+use crate::hash::fnv;
 use crate::structures::finder::Opts;
 use std::collections::HashMap;
